+++ database/src/main/resources/db/migration/V1_43__Update_tx_event_attr_part_2.sql
@@ -1,0 +1,6 @@
+SELECT 'Update tx_msg_event_attr.attr_hash' AS comment;
+-- CREATE EXTENSION IF NOT EXISTS pgcrypto; -- CREATED MANUALLY DUE TO PERMISSIONS ERROR
+
+UPDATE tx_msg_event_attr
+SET attr_hash = encode(digest(concat(attr_idx,attr_key,attr_value)::text, 'sha512'::TEXT), 'base64'::TEXT)
+WHERE attr_hash IS NULL;