+++ database/src/main/resources/db/migration/V1_45__Invalid_denom_cleanup.sql
@@ -1,4 +1,4 @@
 SELECT 'Updating bad denoms' AS comment;
 UPDATE marker_cache
 SET status = 'MARKER_STATUS_UNSPECIFIED'
-WHERE marker_type = 'DENOM'
+WHERE marker_type = 'DENOM';